+++ src/cif.h
@@ -50,6 +50,6 @@ CIFDATA readCifHeader (XFILE * ayb_fp);
 bool writeCIFtoFile ( const CIFDATA  cif, const char * fn, const XFILE_MODE mode);
 bool writeCIFtoStream ( const CIFDATA  cif, XFILE * ayb_fp);
 bool write2CIFfile ( const char * fn, const XFILE_MODE mode, const encInt  intensities, const uint16_t firstcycle, const uint32_t ncycle, const uint32_t ncluster, const uint8_t nbyte);
-CIFDATA spliceCIF(const CIFDATA cif, uint32_t ncycle, uint32_t offset);
+CIFDATA spliceCIF(const CIFDATA cif, uint32_t ncycle, size_t offset);
 
 #endif
